Removed support methods from Home Component
